rename: 실행 클래스 명 변경
