+++ src/main/java/com/likelion/wisesaying/App.java
@@ -1,0 +1,10 @@
+package com.likelion.wisesaying;
+
+import com.likelion.wisesaying.controller.MainController;
+
+public class App {
+    public static void main(String[] args) {
+        MainController app = new MainController();
+        app.run();
+    }
+}
