--- src/main/java/com/likelion/wisesaying/Main.java
@@ -1,8 +1,0 @@
-package com.likelion.wisesaying;
-
-public class Main {
-    public static void main(String[] args) {
-        Application app = new Application();
-        app.run();
-    }
-}
